fix(qualification): repair exact-head CI gates
