+++ test/helpers/release-qualification-gate-fixture.ts
@@ -4,6 +4,7 @@
 import { execFileSync } from "node:child_process";
 import crypto from "node:crypto";
 import fs from "node:fs";
+import path from "node:path";
 
 const REAL_GIT = execFileSync("which", ["git"], { encoding: "utf8" }).trim();
 
@@ -19,6 +20,84 @@ export type ReleaseQualificationFixture = {
   targetSha: string;
   work: string;
 };
+
+export type ReleaseQualificationAuthorityMutation = "gitlink" | "mode" | "symlink";
+
+export function stageReleaseQualificationContract(
+  contractPath: string,
+  contract: Record<string, unknown>,
+  enabled: boolean,
+): void {
+  if (!enabled) return;
+  fs.mkdirSync(path.dirname(contractPath), { recursive: true });
+  fs.writeFileSync(contractPath, `${JSON.stringify(contract, null, 2)}\n`);
+}
+
+export function stageUnlistedQualificationValidatorImport(
+  work: string,
+  validatorPath: string,
+  enabled: boolean,
+): void {
+  if (!enabled) return;
+  const helperPath = path.join(
+    work,
+    "scripts",
+    "checks",
+    "openshell-qualification-unlisted-helper.mts",
+  );
+  fs.writeFileSync(
+    helperPath,
+    [
+      "// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.",
+      "// SPDX-License-Identifier: Apache-2.0",
+      "",
+      "export const unlistedQualificationHelper = true;",
+      "",
+    ].join("\n"),
+  );
+  fs.appendFileSync(validatorPath, '\nimport "./openshell-qualification-unlisted-helper.mts";\n');
+}
+
+export function removeReleaseTagIfPresent(work: string, tag: string): void {
+  try {
+    runReleaseQualificationCommand(work, [
+      "git",
+      "show-ref",
+      "--verify",
+      "--quiet",
+      `refs/tags/${tag}`,
+    ]);
+  } catch {
+    return;
+  }
+  runReleaseQualificationCommand(work, ["git", "tag", "-d", tag]);
+}
+
+export function mutateReleaseQualificationAuthority(
+  fixture: Pick<ReleaseQualificationFixture, "baseSha" | "work">,
+  authorityPath: string,
+  mutation: ReleaseQualificationAuthorityMutation,
+): void {
+  const absolutePath = path.join(fixture.work, authorityPath);
+  if (mutation === "mode") {
+    fs.chmodSync(absolutePath, fs.statSync(absolutePath).mode | 0o100);
+    runReleaseQualificationCommand(fixture.work, ["git", "add", authorityPath]);
+  } else if (mutation === "symlink") {
+    fs.rmSync(absolutePath);
+    fs.symlinkSync("README.md", absolutePath);
+    runReleaseQualificationCommand(fixture.work, ["git", "add", authorityPath]);
+  } else {
+    fs.rmSync(absolutePath);
+    runReleaseQualificationCommand(fixture.work, [
+      "git",
+      "update-index",
+      "--add",
+      "--cacheinfo",
+      `160000,${fixture.baseSha},${authorityPath}`,
+    ]);
+    fs.mkdirSync(absolutePath);
+  }
+}
 
 export function releaseQualificationEnv(
   fixture?: ReleaseQualificationFixture,
